Node labels in TTN
Modified output.c so that the newick string in a .ttn file contains node
labels (the index numbers, same as in .tre file).

This makes it much easier for ttn.R to tell ape which node states go where,
e.g., for plotting tip and node states.

The small_tree.ttn example in the manual now shows node labels.