Absolute ages seem to be working.
Now I have an age-related attribute (atime) for each node.  It inherits
from the ancestor if going left, and it is set to the node creation
time if going right.

Also noted a possible problem with ShowRootState.  Doesn't affect
results but may not be reporting accurately?

And noted that once absolute ages are sorted out for 2regions, should
also implement for 2states.